mysql: add option to force the migration method used [MA-704]
Adds an option to force the desired migration method without doing
fallback to any other methods. In practice, this allows to skip trying
to use replication method and attempt dump method straight away. Also,
setting the method to replication will not fallback to dump, but just
fails if replication is not possible.